requirements.txt
# X (Twitter) Advanced Post & User Scraper
# Production Dependencies

tweepy==4.15.0
httpx==0.28.1
playwright==1.49.1
selectolax==0.3.26
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.1.0
rich==13.9.4
tenacity==9.0.0
fake-useragent==2.0.3
aiofiles==24.1.0
orjson==3.10.14
pydantic==2.10.5
click==8.1.8
tqdm==4.67.1
loguru==0.7.3
